# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agroba Enterprise API Documentation",
        "description": "API documentation for Enterprise Management System",
        "termsOfService": "https://agroba.local/terms",
        "contact": {
            "email": "suyaroff@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://agroba.local",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/api/enterprises": {
            "get": {
                "tags": [
                    "Enterprises"
                ],
                "summary": "Get list of enterprises",
                "operationId": "6e884a4d41b5f90cf27b35b4237c2bae",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "JLB Company"
                                                    },
                                                    "director_name": {
                                                        "type": "string",
                                                        "example": "Suyarov Jamal"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Tashkent, Main street, 32"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "suyaroff@gmail.com"
                                                    },
                                                    "website": {
                                                        "type": "string",
                                                        "example": "jlbcompany.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+998994888515"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-05-19T23:28:08.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-05-19T23:28:08.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://agroba.local/api/enterprises?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://agroba.local/api/enterprises?page=5"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://agroba.local/api/enterprises?page=2"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "&laquo; Previous"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://agroba.local/api/enterprises"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 75
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Enterprises"
                ],
                "summary": "Create a new enterprise",
                "operationId": "1b3fe5edf64cc88b9fc4c7ec413edbcb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "JLB Company"
                                    },
                                    "director_name": {
                                        "type": "string",
                                        "example": "Suyarov Jamal"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Tashkent, Main street, 32"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "suyaroff@gmail.com"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "jlbcompany.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+998994888515"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enterprise created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "JLB Company"
                                                },
                                                "director_name": {
                                                    "type": "string",
                                                    "example": "Suyarov Jamal"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Tashkent, Main street, 32"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "suyaroff@gmail.com"
                                                },
                                                "website": {
                                                    "type": "string",
                                                    "example": "jlbcompany.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+998994888515"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-05-19T23:28:08.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-05-19T23:28:08.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}